% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootMackChainLadder.R
\name{BootMackChainLadder}
\alias{BootMackChainLadder}
\title{Boot Mack Chain Ladder model}
\usage{
BootMackChainLadder(Triangle, B = 100, distNy = "normal", seuil = NA,
  zonnage = FALSE, BF.premiums = NULL, BF.param = c(5, 5), stab = NA)
}
\arguments{
\item{Triangle}{A simple triangle from che Chain-ladder package.}

\item{B}{numeric. The number of bootstrap samples you want}

\item{distNy}{character. Distribution of next-year incremental payments. Either "normal" (default) or "residuals"}

\item{seuil}{numeric. A value of NA (default) will prevent exclusion of residuals, and a numerci value (e.g 2) will exclude all residuals that have an absolution value greater than 2.}

\item{zonnage}{logical. Do you want to force residuals to be resampled inside zonnes ?}

\item{BF.premiums}{If a Bornhuetter-fergusson is needed, input a vector of ultimates premiums here. Otherwise, the BF code will not be triggered.}

\item{BF.param}{A vector of 2 interger that represent (respectively) the number of year of averaging Loss-ratios for the bornhuetter fergusson and then the number of year of applying the bornhuetter fergusson to.}
}
\value{
A BootMackChainLadder object with a lot of information about the bootstrapping. You can plot it, print it and str it to extract information.
}
\description{
This function implement a simple bootstrap of the residuals from the mack model with a one-year reserving risk point of view.
}
\examples{
data(ABC)
BootMackChainLader(Triangle = ABC, B = 100, distNy = "residuals", seuil = 2)
}
