% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootMackChainLadder.R
\name{Corel}
\alias{Corel}
\title{Title}
\usage{
Corel(MBMCL, ...)
}
\arguments{
\item{MBMCL}{A MultiBootMackChainLadder Object}

\item{...}{Elements to be passed to the cor function}
}
\value{
A corelation matrix of next-year IBNRs.
}
\description{
Title
}
\examples{
data(ABC)
triangles <- list(tri1 = ABC, tri2 = ABC, tri3 = ABC)
MBMCL <- MultiBootmackChainLadder(triangles,100)
Corel(MBMCL)
}
