% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootMackChainLadder.R
\name{plot_factors_density}
\alias{plot_factors_density}
\alias{plot_reserve_density}
\alias{plot_ibnr_density}
\alias{plot_burn_in}
\alias{plot_rho}
\alias{plot_resid_norm}
\alias{plot_resid_dens}
\alias{plot_resid_dens2}
\alias{plot_resid_stabilite}
\alias{plot_resid_margins}
\alias{plot_cdr_mw}
\alias{plot_cadences_capi}
\title{plots of EtudePrincipale}
\usage{
plot_factors_density(x)
}
\arguments{
\item{x}{A EtudePrincipale object}
}
\value{
dependns. sometimes thees functions returns a plot.
}
\description{
A serie of plotting functions exists for an etudeprincipale.
 Some more will probably be added later.
}
\details{
Following plotting functions are avaliable : plot_factors_density plot_reserve_density plot_ibnr_density plot_burn_in plot_rho plot_resid_norm plot_resid_dens plot_resid_dens2 plot_resid_stabilite plot_resid_margins plot_cdr_mw plot_cadences_capi
}
