% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootMackChainLadder.R
\name{print.Etudeprincipale}
\alias{print.Etudeprincipale}
\title{print.Etudeprincipale}
\usage{
\method{print}{Etudeprincipale}(x, ...)
}
\arguments{
\item{x}{An EtudePrincipale Object}
}
\value{
returns x after aving printed a lot of informations about it.
}
\description{
print.Etudeprincipale
}
