% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootMackChainLadder.R
\name{Sigma.EtudePrincipale}
\alias{Sigma.EtudePrincipale}
\title{Sigma.EtudePrincipale}
\usage{
Sigma.EtudePrincipale(x)
}
\arguments{
\item{x}{Une étude}
}
\value{
retourne une matrice de corel des CDR du modèle
}
\description{
Sigma.EtudePrincipale
}
