% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootMackChainLadder.R
\name{Analyse}
\alias{Analyse}
\title{Analyse}
\usage{
Analyse(triangles, B = 100, distNy1 = "normal", distNy2 = "normal",
  seuil = NA, pdd = NA, Dossiers = NA, Zonnage.capi = FALSE)
}
\arguments{
\item{triangles}{the set of triangles. Should be a named list with 3 triangles for PSAPs and the last one for PSNEMs.}

\item{B}{Number of bootstrap replicates}

\item{distNy1}{Parameter for the MultiBookMackChainLadder part (psaps) of the procedure.}

\item{distNy2}{Parameter for the BookMackChainLadder part (psnems) of the procedure.}

\item{seuil}{Parameter for both parts.}

\item{pdd}{Should be a data.frame with one row and 3 columns.}

\item{Dossiers}{Working folders informaitons.}

\item{Zonnage.capi}{Should the capitalisation bootstrap be zonned ?}
}
\value{
an "Etudeprincipale" object wich has only a print method and contains everything (try to str it !)
}
\description{
This function perform the analysis from my memoire on a set of triangles.
}
