% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootMackChainLadder.R
\name{BootMackChainLadder}
\alias{BootMackChainLadder}
\title{Boot Mack Chain Ladder model}
\usage{
BootMackChainLadder(Triangle, B = 100, distNy = "normal", seuil = NA,
  zonnage = FALSE, BF.premiums = NULL, BF.param = c(5, 5), stab = NA)
}
\arguments{
\item{Triangle}{A simple triangle from che Chain-ladder package.}

\item{B}{numeric. The number of bootstrap samples you want}

\item{distNy}{character. Distribution of next-year incremental payments. Either "normal" (default) or "residuals"}

\item{seuil}{numeric. Seuil for residuals exclusion. A value of NA (default) will prevent}

\item{zonnage}{logical. Do you want to zonne the residuals ?}

\item{BF.premiums}{les primes pourle BF}

\item{BF.param}{un vectuer de deux entiers représentant le nombre d'année ou la moyenne de LR sera prise pour le BF et le nombre d'année d'application du BF.}
}
\value{
A BootMackChainLadder object with a lot of information about the bootstrapping. You can plot it, print it, str it to extract information.
}
\description{
This function implement a simple bootstrap of the residuals from the mack model with a one-year reserving risk point of view.
}
\examples{
data(ABC)
BootMackChainLader(Triangle = ABC, B = 100, distNy = "residuals", seuil = 2)
}
